Decode some strings to C

+++ include/uv_debug.h
@@ -3,9 +3,9 @@
 
 #include <PR/ultratypes.h>
 
-void func_8023345C(s32, s32, s32);
+void func_8023345C(u8, u8, u8);
 void func_80233590(void);
 void func_803182A0(void);
-void uvDbgCnt(u8, s32);
+void uvDbgCnt(u8, u8);
 
 #endif // UV_DEBUG_H
